ai-service/requirements.txt
fastapi
uvicorn
sentence-transformers
langchain
langchain-groq
pydantic
python-dotenv
